Move `#details` to package manager specific lockfile class
