+++ npm_and_yarn/lib/dependabot/npm_and_yarn/file_parser/yarn_lock.rb
@@ -25,6 +25,22 @@ module Dependabot
             raise Dependabot::DependencyFileNotParseable, @dependency_file.path
           end
         end
+
+        def details(dependency_name, requirement, _manifest_name)
+          details_candidates =
+            parsed.
+            select { |k, _| k.split(/(?<=\w)\@/)[0] == dependency_name }
+
+          # If there's only one entry for this dependency, use it, even if
+          # the requirement in the lockfile doesn't match
+          if details_candidates.one?
+            details_candidates.first.last
+          else
+            details_candidates.find do |k, _|
+              k.scan(/(?<=\w)\@(?:npm:)?([^\s,]+)/).flatten.include?(requirement)
+            end&.last
+          end
+        end
       end
     end
   end
